Adds tool configuration via JSON file
Introduces a configuration file to enable/disable tools.
This allows users to control which tools are displayed and used.

The configuration file is automatically generated/updated with all discovered tools.
It is loaded at startup and used to filter the list of available tools.
A new config is created if one doesn't already exist.